+++ inst/doc/stickers.R
@@ -1,10 +1,12 @@
 ## ----setup, include = FALSE---------------------------------------------------
 source("https://raw.githubusercontent.com/Flavjack/inti/master/pkgdown/favicon/docs.r")
 
-## ----echo = TRUE, eval = T----------------------------------------------------
+## ----echo = TRUE--------------------------------------------------------------
 library(huito)
 
-huito_fonts()
+font <- "Permanent Marker"
+
+huito_fonts(font)
 
 label <- label_layout(size = c(5.08, 5.08)
                       , border_color = NA
@@ -14,14 +16,14 @@ label <- label_layout(size = c(5.08, 5.08)
                 , size = c(7, 7)
                 , position = c(2.55, 2.52)
                 ) %>%
-  include_text(value = "H", size = 45 , position = c(1.15, 2.7), color = "#00a85a") %>%
-  include_text(value = "u", size = 45 , position = c(2.07, 2.7), color = "#f58735") %>%
-  include_text(value = "i", size = 45 , position = c(2.73, 2.7), color = "#4774b8") %>%
+  include_text(value = "H", size = 45 , position = c(1.15, 2.7), color = "#00a85a", font) %>%
+  include_text(value = "u", size = 45 , position = c(2.07, 2.7), color = "#f58735", font) %>%
+  include_text(value = "i", size = 45 , position = c(2.73, 2.7), color = "#4774b8", font) %>%
   include_image(value = "https://flavjack.github.io/huito/img/huito_fruit.png"
                 , size = c(1.3, 1.3) 
                 , position = c(4.06, 2.6)
                 ) %>%
-  include_text(value = "t", size = 45 , position = c(3.33, 2.7), color = "#a9518b") %>%
+  include_text(value = "t", size = 45 , position = c(3.33, 2.7), color = "#a9518b", font) %>%
   include_shape(size = 4.1
                 , border_width = 3
                 , border_color = "black"
@@ -38,7 +40,7 @@ label <- label_layout(size = c(5.08, 5.08)
 
 ## ---- echo = TRUE-------------------------------------------------------------
 label %>% 
-  label_print(mode = "preview", smpres = 300, viewer = F)
+  label_print(mode = "preview")
 
 ## ---- echo = TRUE-------------------------------------------------------------
 sticker <- label %>%